[Added] Start of CLI feature
Commands can be added to the commandList with a list of arguments. Arguments can be given a dependant tag meaning their possible value changes according to the previous argument value. Code checks if command exists, if correct amount of arguments are given, if argument values given are correct and runs the command function if these all pass.